backend/requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
requests
pydantic
PyPDF2
python-docx
nltk
